[#66093] Add general error for missing attributes

+++ app/models/acts_as_customizable/calculated_value.rb
@@ -48,14 +48,14 @@ module ActsAsCustomizable::CalculatedValue
       end
 
       enabled_ids = enabled_custom_field_ids
-
+      given = calculated_value_fields_given(custom_fields:, enabled_ids:)
       result = calculate_custom_fields_result(
-        given: calculated_value_fields_given(custom_fields:, enabled_ids:),
+        given:,
         to_compute: calculated_value_fields_to_compute(custom_fields:, enabled_ids:)
       )
 
       self.custom_field_values = custom_fields.to_h { [it.id, result[it.column_name]] }
-      refresh_calculation_errors!(result)
+      refresh_calculation_errors!(given, custom_fields.map(&:id), result)
     end
 
     private
@@ -86,7 +86,7 @@ module ActsAsCustomizable::CalculatedValue
       cf_id.sub("cf_", "").to_i
     end
 
-    def refresh_calculation_errors!(result)
+    def refresh_calculation_errors!(referenced_values, calculated_field_ids, result)
       return unless is_a?(Project)
 
       failed_calculations, successful_calculations = result.partition do |_, calculation_result|
@@ -95,6 +95,8 @@ module ActsAsCustomizable::CalculatedValue
 
       successfully_calculated_cfs = successful_calculations.map { |cf_id, _| to_id(cf_id) }
       remove_calculated_value_errors(successfully_calculated_cfs)
+
+      create_errors_for_missing_attributes(referenced_values, calculated_field_ids)
 
       failed_calculations.to_h.each_key do |cf_id|
         create_calculated_value_error(to_id(cf_id), "ERROR_MATHEMATICAL")
@@ -105,7 +107,19 @@ module ActsAsCustomizable::CalculatedValue
       CalculatedValueError.create(project: self, custom_field_id:, error_code:)
     end
 
+    def create_errors_for_missing_attributes(referenced_values, calculated_field_ids)
+      cf_ids_with_missing_values = referenced_values.filter_map { |k, v| to_id(k) if v.nil? }
+      cf_ids_with_missing_values.each do |_missing_cf_id|
+        calculated_field_ids.each do |calculated_cf_id|
+          # TODO: insert missing custom field name into the error message:
+          create_calculated_value_error(calculated_cf_id, "ERROR_MISSING_VALUE")
+        end
+      end
+    end
+
     def remove_calculated_value_errors(custom_field_ids = [])
+      return if custom_field_ids.empty?
+
       CalculatedValueError.where(project: self, custom_field_id: custom_field_ids).delete_all
     end
   end
